[http]
Support all HTTP verbs: Server.route() and Server.findCallback() take a HTTP (method) parameter representing the verb. (It can also be parsed with Server.methodType()). Adding a form to a template to allow sending of POST request. Add exception handling to handleCallback() - we ought to serve a 501 if an exception is thrown...