[http]
Adding a number of (empty) modules for use in the forum.exe: BbCode, HtmlSanitizer, OEmbed, Textile (all under httplib.mod).

These reference third-party dlls which have been added to contrib/server.
The intent is to allow use of these components as optional 'plugins' for the forum when it's up and running...