[http]
A better attempt at user authentication after login - generate a session token for the user.
Take 60 bytes of cryptographic randomness and base64 encode it.
Send it as a cookie to the user, and store in a Sessions DB table.
When the user makes a request, if the cookie matches an entry in the table, then assume that's the user.
Session tokens will expire after 30 mins.