[http]
Improved algorithm for sending <textarea> updates to myserver.
Sends two positions and some replacement text (which can be empty): (0,0, "") resets the string on the server.
Optimized cases for single char adds/deletes, more general test for other kinds of updates.
(There are other potential optimizations which could be made, such as caching the last update position so we know where to expect the next change, but these are quality-of-implementation issues...)