[http]
Adding an extension method for System.Exception: ex.GetExtendedStackTrace()
This method tries to return location data for the exception's stack trace, in a standard, tool-friendly "file:line:column:" format.
It only works if .pdb files are generated for all the built assemblies, so (finally) splitting up the build into seperate Debug and Release configurations.

For a Debug build: 'make DEBUG=1' or just 'make' (it's the default)
For a Release build: 'make RELEASE=1' or 'make DEBUG=0'

Assemblies are output into '$(BIN)/debug/' or '$(BIN)/release/'.
Preprocessor symbols DEBUG or RELEASE are #defined appropriately (these are apparently the standards for .NET/C# projects).
Debug build also specifies ncc -g to enable debug symbols (pdbs).
Release build also specifies ncc -O for optimized code.

Added a make rule for targets matching %-full: build % in Debug and then Release. This is pretty handy for keeping debug/release builds in sync. (And also for running tests in both configs, when we get to that point...)

Had to change all paths to assets currently using "../src" to "../../src" instead, but this is horrible anyway, a better solution is needed.