[http]
Semi-horrendous signin/singnout flow.
We now have three kinds of session tokens: secure, insecure and temporary:
Secure tokens are valid for 30 days and have the Secure flag - can only be used by the nav over https.
Insecure tokens are valid for 30 minutes are can be used by the main window over http.
Temporary tokens are valid for 1 minute and are used to coordinate the two windows.

At sign in, the following happens:
 - the username/password are validated and a secure session is stored in DB and as a cookie (_session).
 - the nav causes a 303 redirect back to /
 - the main page generates a temporary token with user ID 0 and posts it to the nav frame (tokTmp).
 - the nav frame makes a secure post to /tok-set with the anonymous temporary token
 - The server verifies the temporary token, and uses the nav's _session token to get the UserID
 - the server generates a second temporary token with the UserID and sends it to the nav.
 - the nav posts a message back to the main window with the second token (tokUp)
 - the main window posts the temporary token back to the server
 - server gets the UserID, generates an insecure session token (_isession) and sends as a cookie to the main window.
 - Main window is now logged in with the insecure token: force a reload so that the cookie value can be read.
 - phew.

It ought to be possible to opmitize this flow somewhat - there are at least two redirects and a location.reload() at the moment...

- Sign out is a little easier, we just trash every token with the UserID from the database and expire the _session and _isession tokens.
- We don't need to reload, we just pretend we didn't see _isession during the response where we expire it.

Next stage of the process is to try to pull core pieces of functionality out into base handlers which can be shared between different pages...