[http]
Start to split functionality of ForumServer out into Handlers.
Helps to reduce the amount of repetition in the code and make the underlying structure clearer.
forum.BaseHandler is intended to be the root of all forum handlers.
A test version of the index is available at /index, and all of the /nav-{{action}}s are handled by instances of NavHandler.

Added IHandlerFactory interface: makeHandler() HandlerData -> Handler, Server.route() can take instances of IHanderFactory.
CallbackHandlerWrapper.unwrap() is now a called CallbackHandlerFactory.makeHandler()

Makefile modified so that the src/ parameter to $(emit_rule) is now a list of folders.