[http]
Getting rid of the location.reload() in the signin flow, use an ajax POST to update the window instead.
Slightly changed the internals of httplib.ComplexHandler.callback(), adding a renderPage() method.
In forum.BaseHandler, we overload renderPage() to allow a call to page.renderContent() instead of page.render() if the variable _$content is present in the querystring or the post body.
Template variables RenderFullPage vs RenderContent are used to control the page rendering.
The 'main' <div> container of the <body> has an id of #mainspace. In a content render, we render only its direct contents, so we can use $().html() to update it in real time, just need to make a few JS calls post-update...
This is all in the template.nyml, so ought to work on any page.
Moved some login flow code into a method in the BaseHandler (temporaryLoginHandling()) for the time being, as the old indexResponse() method is broken now anyway.
Still some untidiness and lots of potential for improvement - this ajax technique ought to let us remove a few redirects as well...