[http]
Adding 'create new thread' functionality, available from the category page.
The thread title can be created and optionally the first post in the thread.
Stubbed out renderer selection and preview, but these are non-functional for the time being.
Added CSS for 'expanding-area': a combination <textarea>/<pre> component which expands to fit user input as it is typed.
Currenly available in 2 variants: expanding-h1-area for editable titles, expanding-p-area for posts.
A small amount of jQuery is needed to make it work, that's in newthread.js for the time being.

Supporting core tech tweaks:
* httplib.Database.withTransaction() can now take a second lambda to handle/report exceptions thrown during the tranaction.
* httplib.Server.processPostInputStream() transforms '+' chars to ' ' prior to URL decoding if the content type is "x-www-form-urlencoded"...