[http]
We can use nustache tempalates to generate CSS data:
A file called custom-styles.css?c=x360, where c is the colour ID.
It also checks the isession token for userid and gets the user's font profile.
Returned CSS has all the fonts and colours needed for a page, this means that when setFont() and setHdrCol() are called, the data is 90% there already.
So now we don't 'pop' from monochrome Arial into the page layout.
The setFont()/setHdrCol() still work as before (for the user page etc.)

All the font map data is now in data and written out as part of the font data, both as JavaScript (for client) and YAML (for server).
The font map syntax is deliberately chosen not to look like CSS. (It's a simple line-based awk format)