[http]
Optimized font loading - store the css file for each font as 'import' in the fontdata.
Add selected @imports to the custom-styles.css.
We're only loading 3 fonts per page rather than all 8.
For the user/prefs page, we load all the fonts via jQuery (and instruct the nav to do so as well),
so we can see the result of font switching live on the page, but the load doesn't block the initial page load.