[http]
Fix use of location.reload() in thread page upon submission of new reply.
We make an ajax call to reload the page content(with ?_$content=1), same as for a token check.
That ajax call is now in a function called reloadPageContent() for future use.
So now, new posts seamlessly appear within the page - nice!
(I think we will track replies to a post after the post, so if a reply ends up on a different page, there will still be a visual change to the page...)
Had to remove Hyphenator.js as it was too slow - causing page loads to take forever.
Replaced it which hypher jquery plugin, which seems much much faster.
At the moment it's hyphenating $('*') which is probably excessive, but is still acceptable.