[http]
Use ""s rather than ''s to conatain expanded Nustache objects in HTML/JS fragments.
Nustache escapes " into &quot; but doesn't escape a ', so it was possible for user input to cause syntax errors on the page.
We're still forced to use '{{id}}' and '#{{Number}}' in HTML fragments inside ""s, but we're just using ids here, so this _should_ be OK.