[http]
Marginally better signin/signout behaviour on user page.
Change signout=1 on QS to _$signout=1, so it (eventually) gets stripped from the URL.
Add a _initHook variable which can point to a function called from pageInit()
Move prefs.js drag/drop handling into it's own function (so it can be called via _initHook).
Add error handling to drag/drop so it won't execute if the required <div>s are not present in the page.
TODO: revisit signin/signout in general...