[http]
"Better" handling of font changes on signin/signout - maybe 95% working, most of the time.
Moved the initialization of the _g object into a function (resetProperties()) the div#mainspace.
So we can update these values when a _$content=1 reload takes place.
Also added a (rather hacky) way to force a reload onto custom-styles.css.
(To consider: we _could_ render user pages in the font which that user has selected, although it might be perceived as buggy/inconsistent...)