[http]
Some work on quote behaviour in thread pages.
Style <q> the same as <blockquote> so it can be (ab)used to wrap quoted text.
Style <cite> inside <q> as an italic block with a trailing ':' (to mark the quote user/author)
Add data-author='' attribute to div#content, so it can be extracted programatically when quoting.