[http]
First pass outline thumbs/bookmarks functionality.
Added two new tables (Thumbs and Bookmarks) with UserId/PostId columns.
When a thread is rendered, we can check for thumbs/bookmarks for each post.
We note the total for each post, and if the signed in user has either.
We can use this info to render slightly differently the post.
A couple of ajax endpoints to toggle thumbs/bookmarks on/off for the signed in user.
Need to delay call to initThreadPage until $(window).load() as it depends on _g.ColourId initialized in initPage() -- this may need a revisit...