[http]
Move page Scripts data out of partials into top-level array (same as Stylesheets).
When pages are rendered, a partial {{>_Scripts}} can be used to indicate where scipts should be inserted.
(This allows us to place them in the page body, but have some additional script code placed afterwards if desired.)
Page now has a renderFileElementList() function which expands lists of files into a series of HTML elements (with optional additional attributes).
This should make it easier to build concatenated JS or CSS files per-page, to reduce the number of files downloaded (it should also make it easier to insert minimization as well).
Also attempt to normalize script partials to CommmonScript/PageScript for both nav and mainpage.