[http]
Tidy up OEmbed a little.
(We may want to rename this module, as it's really a "bare-URL transformer", not
just OEmbed.)
Allow a matching provider to 'early out' of the provider search; preventing
providers from later in the chain from matching.
Added a DefaultUrlProvider, which builds <a href="URL>URL</a> from a bare URL,
with a suitably broad pattern at the end of the provider list.