[http]
ServerConfiguration.StripPortFromDefaultPrefixes flag, which will allow removal
of port information when constructing the default prefixes.
This could be useful if the server is running in a redirected environment
(eg :80 -> :8080 and :443 -> :8443) and we want to specify the ports that the
server actually listens on, but a portless URL is a valid way to address the
server, so we want to build portless URLs into application pages.