[http]
Added compiler flags for DOTNET and MONO depending on platform.
Programmatically force use of TLS 1.0 for SSL connections on Mono. This is the
strongest encryption we can currently enforce until TLS 1.1/1.2 become available
on Mono.
(Windows users will have to solve this externally with configuration changes at
the HTTP.SYS level, possibly with use of a third-party tool. See for example:
https://www.nartac.com/Products/IISCrypto)