[http]
Sign out confirmation.
Clicking the "Sign out" button posts a message to main window which opens an
OK/Cancel dialogue box. If the user hits yes, we go ahead and tell the nav to
submit the form, which causes the sign out to happen.
Also some CSS to cover fonts in the dialogs, and custom dialogue code (allows
use of Esc or Space/Enter keyboard shortcuts when interacting with the
dialogue).