[http]
(Start of) Admin page.
For now, we really just want an "at a glance" idea of how long the server has
been up, and if any errors happened (i.e. exceptions thrown) recently.
This page is only viewable over HTTPS by users with the IsAdmin flag set to TRUE
in the Users table. Any other attempts to view it will trigger a 404 response.
It displays the server uptime and a "Live Error Log", which updates in real
time, every 10 seconds. (The error log is effected by adding a new LogTarget
which writes Errors & Fatals to "forum-live.log", and monitoring this specific
file.)