[http]
Category selection on newthread page - use a <select>/select2.
Use ajax to get category list from server as JSON  from POST /categories.
/th/new available on both GET and POST - if called from the category page, then
the relevant category is selected as the default.
CreateThreadHandler is passed the category name instead of the id, it can call
Server.getCategoryId() to get the catid which is needed.