[http]
Move 'consolidation' logic into its own class (Consolidator). Also move some
stuff out of page.n that belongs with it. We now check to see if any of the
input files are 'younger' than output before deciding to rebuild the output file
-- we also store the set of files produced for each output, so if we _don't_
rebuild the output, we know what files to serve. This is better than before when
we would unconditionally build the consolidated output file every time. It
should also be easier to add minification to Consolidator class now. When server
starts, we make sure the cache folder exists and is empty (based on value of
Configuration.Consolidation.CachePath).